requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
discord.py==2.3.2
httpx>=0.24.0,<0.25.0
python-multipart==0.0.6
supabase==2.0.3
openai==1.3.0
pydantic==2.4.2
pytest==7.4.3 